--- lib/main2.dart
@@ -1,19 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:namma_metro/loading_page.dart';
-
-void main() => runApp(const MyApp());
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Loading Page Exmple',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: const LoadingPage(),
-    );
-  }
-}
